refactor(search): reduce file size by extracting components
