+++ src/search/components/SearchResults/SearchResults.tsx
@@ -1,0 +1,50 @@
+import React, { Fragment } from 'react';
+
+import { Blankslate, Container, Pagination, Spacer, Spinner } from '@viaa/avo2-components';
+
+import SearchResultItem from './SearchResultItem';
+import { SearchResultsProps } from './types';
+
+const SearchResults = ({
+	currentPage,
+	loading,
+	data,
+	pageCount,
+	setPage,
+	...resultProps
+}: SearchResultsProps) => {
+	return (
+		<Container mode="vertical">
+			<Container mode="horizontal">
+				{loading ? (
+					<div className="o-flex o-flex--horizontal-center">
+						<Spinner size="large" />
+					</div>
+				) : data && data.count !== 0 ? (
+					<Fragment>
+						<ul className="c-search-result-list">
+							{data.results.map((result, index) => (
+								<SearchResultItem
+									{...resultProps}
+									key={`search-result-item-${index}`}
+									result={result}
+								/>
+							))}
+						</ul>
+						<Spacer margin="large">
+							<Pagination pageCount={pageCount} currentPage={currentPage} onPageChange={setPage} />
+						</Spacer>
+					</Fragment>
+				) : (
+					<Blankslate
+						body=""
+						icon="search"
+						title="Er zijn geen zoekresultaten die voldoen aan uw filters."
+					/>
+				)}
+			</Container>
+		</Container>
+	);
+};
+
+export default SearchResults;
