+++ src/search/components/index.ts
@@ -1,0 +1,4 @@
+import SearchFilterControls from './SearchFilters/SearchFilterControls';
+import SearchResults from './SearchResults/SearchResults';
+
+export { SearchFilterControls, SearchResults };
