+++ src/search/views/Search.tsx
@@ -1,5 +1,4 @@
 import {
-	Blankslate,
 	Button,
 	Container,
 	Dropdown,
@@ -8,16 +7,9 @@ import {
 	Form,
 	FormGroup,
 	Navbar,
-	Pagination,
-	SearchResult,
-	SearchResultSubtitle,
-	SearchResultThumbnail,
-	SearchResultTitle,
 	Select,
 	Spacer,
-	Spinner,
 	TextInput,
-	Thumbnail,
 	Toolbar,
 	ToolbarItem,
 	ToolbarLeft,
@@ -27,7 +19,6 @@ import {
 } from '@viaa/avo2-components';
 import { Avo } from '@viaa/avo2-types';
 import {
-	capitalize,
 	cloneDeep,
 	compact,
 	every,
@@ -40,50 +31,18 @@ import {
 	pickBy,
 } from 'lodash-es';
 import queryString from 'query-string';
-import React, { Fragment, FunctionComponent, ReactNode, useEffect, useState } from 'react';
+import React, { Fragment, FunctionComponent, useEffect, useState } from 'react';
 import { connect } from 'react-redux';
-import { RouteComponentProps } from 'react-router';
-import { Link } from 'react-router-dom';
 import { Dispatch } from 'redux';
 
 import { getSearchResults } from '../store/actions';
 import { selectSearchLoading, selectSearchResults } from '../store/selectors';
 
-import {
-	CheckboxDropdownModal,
-	CheckboxOption,
-} from '../../shared/components/CheckboxDropdownModal/CheckboxDropdownModal';
-import { DateRangeDropdown } from '../../shared/components/DateRangeDropdown/DateRangeDropdown';
 import { copyToClipboard } from '../../shared/helpers/clipboard';
-import { formatDate } from '../../shared/helpers/formatters/date';
-import { formatDuration } from '../../shared/helpers/formatters/duration';
-import { generateSearchLink } from '../../shared/helpers/generateLink';
-import { LANGUAGES } from '../../shared/helpers/languages';
-
-interface SearchProps extends RouteComponentProps {
-	searchResults: Avo.Search.Response | null;
-	searchResultsLoading: boolean;
-	search: (
-		orderProperty: Avo.Search.OrderProperty,
-		orderDirection: Avo.Search.OrderDirection,
-		from: number,
-		size: number,
-		filters?: Partial<Avo.Search.Filters>,
-		filterOptionSearch?: Partial<Avo.Search.FilterOption>
-	) => Dispatch;
-}
+import { SearchFilterControls, SearchResults } from '../components';
+import { SearchFilterFieldValues, SearchFilterMultiOptions, SearchProps, SortOrder } from './types';
 
 const ITEMS_PER_PAGE = 10;
-
-interface SortOrder {
-	orderProperty: Avo.Search.OrderProperty;
-	orderDirection: Avo.Search.OrderDirection;
-}
-
-interface SearchResults {
-	count: number;
-	items: Avo.Search.ResultItem[];
-}
 
 const DEFAULT_FORM_STATE: Avo.Search.Filters = {
 	query: '',
@@ -104,7 +63,7 @@ const DEFAULT_FORM_STATE: Avo.Search.Filters = {
 const DEFAULT_SORT_ORDER: SortOrder = {
 	orderProperty: 'relevance',
 	orderDirection: 'desc',
-} as SortOrder;
+};
 
 const Search: FunctionComponent<SearchProps> = ({
 	searchResults,
@@ -117,9 +76,7 @@ const Search: FunctionComponent<SearchProps> = ({
 	const [sortOrder, setSortOrder]: [SortOrder, (sortOrder: SortOrder) => void] = useState(
 		DEFAULT_SORT_ORDER
 	);
-	const [multiOptions, setMultiOptions] = useState({} as {
-		[key: string]: Avo.Search.OptionProp[];
-	});
+	const [multiOptions, setMultiOptions] = useState({} as SearchFilterMultiOptions);
 	const [currentPage, setCurrentPage] = useState(0);
 	const [searchTerms, setSearchTerms] = useState('');
 	const [isOptionsMenuOpen, setIsOptionsMenuOpen] = useState(false);
@@ -227,16 +184,8 @@ const Search: FunctionComponent<SearchProps> = ({
 	// Only execute this effect once after the first render (componentDidMount)
 	useEffect(getFiltersFromQueryParams, []);
 
-	// handleFilterOptionSearchChange = (event: ChangeEvent) => {
-	// 	const target = event.target as HTMLInputElement;
-	// 	if (target) {
-	// 		const { name, value } = target;
-	// 		setDeepState(this, `filterOptionSearch.${name}`, value).then(noop);
-	// 	}
-	// };
-
 	const handleFilterFieldChange = async (
-		value: string | string[] | Avo.Search.DateRange | null,
+		value: SearchFilterFieldValues,
 		id: Avo.Search.FilterProp
 	) => {
 		if (value) {
@@ -257,12 +206,9 @@ const Search: FunctionComponent<SearchProps> = ({
 
 	const handleOrderChanged = async (value: string = 'relevance_desc') => {
 		const valueParts: string[] = value.split('_');
-		const orderProperty: Avo.Search.OrderProperty = valueParts[0] as Avo.Search.OrderProperty;
-		const orderDirection: Avo.Search.OrderDirection = valueParts[1] as Avo.Search.OrderDirection;
-		setSortOrder({
-			orderProperty: orderProperty as Avo.Search.OrderProperty,
-			orderDirection: orderDirection as Avo.Search.OrderDirection,
-		});
+		const orderProperty = valueParts[0] as Avo.Search.OrderProperty;
+		const orderDirection = valueParts[1] as Avo.Search.OrderDirection;
+		setSortOrder({ orderProperty, orderDirection });
 
 		// Reset to page 1 when search is triggered
 		setCurrentPage(0);
@@ -286,180 +232,10 @@ const Search: FunctionComponent<SearchProps> = ({
 		});
 	};
 
-	const renderCheckboxDropdownModal = (
-		label: string,
-		propertyName: Avo.Search.FilterProp,
-		disabled: boolean = false,
-		style: any = {}
-	): ReactNode => {
-		const checkboxMultiOptions = (multiOptions[propertyName] || []).map(
-			(option: Avo.Search.OptionProp): CheckboxOption => {
-				let label = capitalize(option.option_name);
-				if (propertyName === 'language') {
-					label = languageCodeToLabel(option.option_name);
-				}
-				return {
-					label,
-					optionCount: option.option_count,
-					id: option.option_name,
-					checked: ((formState[propertyName] as string[]) || []).includes(option.option_name),
-				};
-			}
-		);
-
-		return (
-			<li style={{ display: 'flex', ...style }}>
-				<CheckboxDropdownModal
-					label={label}
-					id={propertyName as string}
-					options={checkboxMultiOptions}
-					disabled={disabled}
-					onChange={async (values: string[]) => {
-						await handleFilterFieldChange(values, propertyName);
-					}}
-				/>
-			</li>
-		);
-	};
-
-	const languageCodeToLabel = (code: string): string => {
-		return capitalize(LANGUAGES.nl[code]) || code;
-	};
-
-	const renderDateRangeDropdown = (
-		label: string,
-		propertyName: Avo.Search.FilterProp
-	): ReactNode => {
-		const range: Avo.Search.DateRange = get(formState, 'broadcastDate') || { gte: '', lte: '' };
-		range.gte = range.gte || '';
-		range.lte = range.lte || '';
-
-		return (
-			<li style={{ display: 'flex' }}>
-				<DateRangeDropdown
-					label={label}
-					id={propertyName}
-					range={range as { gte: string; lte: string }}
-					onChange={async (range: Avo.Search.DateRange) => {
-						await handleFilterFieldChange(range, propertyName);
-					}}
-				/>
-			</li>
-		);
-	};
-
-	const renderFilterControls = () => {
-		return (
-			<div className="c-filter-dropdown-list">
-				{renderCheckboxDropdownModal('Type', 'type')}
-				{renderCheckboxDropdownModal('Onderwijsniveau', 'educationLevel')}
-				{renderCheckboxDropdownModal('Domein', 'domain', true)}
-				{renderCheckboxDropdownModal('Vak', 'subject')}
-				{renderCheckboxDropdownModal('Trefwoord', 'keyword')}
-				{renderCheckboxDropdownModal('Serie', 'serie')}
-				{renderDateRangeDropdown('Uitzenddatum', 'broadcastDate')}
-				{renderCheckboxDropdownModal('Taal', 'language')}
-				{renderCheckboxDropdownModal('Aanbieder', 'provider', false, { marginRight: 0 })}
-			</div>
-		);
-	};
-
-	const hasFilters = () => {
-		return !isEqual(formState, DEFAULT_FORM_STATE);
-	};
-
 	const deleteAllFilters = () => {
 		setFormState({
 			...DEFAULT_FORM_STATE,
 		});
-	};
-
-	const renderSearchResult = (result: Avo.Search.ResultItem) => {
-		const metaData = [];
-		let thumbnailMeta = '';
-		if (result.administrative_type === 'audio' || result.administrative_type === 'video') {
-			thumbnailMeta = formatDuration(result.duration_seconds || 0);
-			metaData.push({
-				label: thumbnailMeta,
-			});
-		} else {
-			// TODO get number of items from result item after bart updates the elasticsearch index
-			thumbnailMeta = `${25} items`;
-			metaData.push({
-				label: thumbnailMeta,
-			});
-		}
-		const contentLink = `/item/${result.id}`;
-
-		return (
-			<SearchResult
-				key={`search-result-${result.id}`}
-				type={result.administrative_type}
-				date={formatDate(result.dcterms_issued)}
-				tags={[
-					{ label: 'Redactiekeuze', id: 'redactiekeuze' },
-					{ label: 'Partner', id: 'partner' },
-				]}
-				viewCount={412}
-				bookmarkCount={85}
-				// duration={formatDuration(result.duration_seconds || 0)}
-				description={result.dcterms_abstract}
-				onToggleBookmark={(active: boolean) => handleBookmarkToggle(result.id, active)}
-			>
-				<SearchResultTitle>
-					<Link to={contentLink}>{result.dc_title}</Link>
-				</SearchResultTitle>
-				<SearchResultSubtitle>
-					{generateSearchLink('provider', result.original_cp, 'c-body-2', () =>
-						handleOriginalCpLinkClicked(result.id, result.original_cp)
-					)}
-				</SearchResultSubtitle>
-				<SearchResultThumbnail>
-					<Link to={contentLink}>
-						<Thumbnail
-							category={result.administrative_type as any}
-							src={result.thumbnail_path}
-							label={result.administrative_type}
-						/>
-					</Link>
-				</SearchResultThumbnail>
-			</SearchResult>
-		);
-	};
-
-	const renderSearchResults = () => {
-		return (
-			<Container mode="vertical">
-				<Container mode="horizontal">
-					{!searchResultsLoading && searchResults && searchResults.count !== 0 && (
-						<Fragment>
-							<ul className="c-search-result-list">
-								{searchResults.results.map(renderSearchResult)}
-							</ul>
-							<Spacer margin="large">
-								<Pagination
-									pageCount={pageCount}
-									currentPage={currentPage}
-									onPageChange={setPage}
-								/>
-							</Spacer>
-						</Fragment>
-					)}
-					{!searchResultsLoading && searchResults && searchResults.count === 0 && (
-						<Blankslate
-							body=""
-							icon="search"
-							title="Er zijn geen zoekresultaten die voldoen aan uw filters."
-						/>
-					)}
-					{searchResultsLoading && (
-						<div className="o-flex o-flex--horizontal-center">
-							<Spinner size="large" />
-						</div>
-					)}
-				</Container>
-			</Container>
-		);
 	};
 
 	const setPage = async (pageIndex: number): Promise<void> => {
@@ -508,6 +284,7 @@ const Search: FunctionComponent<SearchProps> = ({
 	];
 	const defaultOrder = `${sortOrder.orderProperty || 'relevance'}_${sortOrder.orderDirection ||
 		'desc'}`;
+	const hasFilters = !isEqual(formState, DEFAULT_FORM_STATE);
 	const resultsCount = get(searchResults, 'count', 0);
 	// elasticsearch can only handle 10000 results efficiently
 	const pageCount = Math.ceil(Math.min(resultsCount, 10000) / ITEMS_PER_PAGE);
@@ -603,7 +380,7 @@ const Search: FunctionComponent<SearchProps> = ({
 									<FormGroup inlineMode="shrink">
 										<Button label="Zoeken" type="primary" onClick={copySearchTermsToFormState} />
 									</FormGroup>
-									{hasFilters() && (
+									{hasFilters && (
 										<FormGroup inlineMode="shrink">
 											<Button
 												label="Verwijder alle filters"
@@ -615,11 +392,23 @@ const Search: FunctionComponent<SearchProps> = ({
 								</Form>
 							</div>
 						</Spacer>
-						{renderFilterControls()}
+						<SearchFilterControls
+							formState={formState}
+							handleFilterFieldChange={handleFilterFieldChange}
+							multiOptions={multiOptions}
+						/>
 					</Spacer>
 				</Container>
 			</Navbar>
-			{renderSearchResults()}
+			<SearchResults
+				currentPage={currentPage}
+				data={searchResults}
+				handleBookmarkToggle={handleBookmarkToggle}
+				handleOriginalCpLinkClicked={handleOriginalCpLinkClicked}
+				loading={searchResultsLoading}
+				pageCount={pageCount}
+				setPage={setPage}
+			/>
 		</Container>
 	);
 };
